+++ src/app/SignOutBtn.tsx
@@ -2,7 +2,7 @@
 
 import { signOut } from "firebase/auth";
 import { useRouter } from "next/navigation";
-import { useAuth } from "./context/authContext";
+import { useAuth } from "./authContext";
 
 const SignOutBtn = () => {
   const { auth } = useAuth();
